requirements.txt
python-binance
python-dotenv
click
colorama
pandas
flask
